requirements.txt
bcrypt==3.2.0
blinker==1.4
cffi==1.14.5
click==7.1.2
dnspython==2.1.0
email-validator==1.1.2
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
greenlet==1.0.0
idna==3.1
importlib-metadata==4.0.1
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
Pillow==8.2.0
pkg-resources==0.0.0
pycparser==2.20
six==1.15.0
SQLAlchemy==1.4.11
typing-extensions==3.7.4.3
Werkzeug==1.0.1
WTForms==2.3.3
zipp==3.4.1
python-dotenv==0.20.0